Refactor prototype single-file picking functionality, feature #45
Refactored BrowserFragment into AbstractBrowserFragment, which
BrowserFragment and a new PickerFragment both extend.

Refactored BrowserActivity functional tests with an abstract class
holding common tests and helper methods, which specific Activity
functional tests extend.  Added a PickerActivityFunctionalTest.

Base tests pass, file picking works.

Need to add multiple file picking and clean up for production.
