+++ explorer/build.gradle
@@ -49,7 +49,7 @@ dependencies {
     compile 'com.melnykov:floatingactionbutton:1.3.0'
 
     androidTestCompile ('com.android.support.test:runner:0.2') {
-        // Testing support library depends on old support-annotations versions
+        // Testing support library depends on old support-annotations version
         exclude module: 'support-annotations'
     }
     androidTestCompile ('com.android.support.test.espresso:espresso-core:2.1') {
